feat: STEP2 CRUD 구현

+++ Notes/Model/NoteModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct NoteModel: Codable {
     var title: String
     var body: String
-    var lastModified: Int
+    var lastModified: Double
     
     enum CodingKeys: String, CodingKey {
         case title, body
